# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Campus GatePass System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /gate-passes:
    get:
      operationId: listGatePasses
      tags: [gate-passes]
      summary: List gate passes (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected, out, returned]
        - name: studentId
          in: query
          schema:
            type: integer
        - name: hostelId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of gate passes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePassListResponse"
    post:
      operationId: createGatePass
      tags: [gate-passes]
      summary: Apply for a gate pass (student)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GatePassInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePass"
  /gate-passes/{id}:
    get:
      operationId: getGatePass
      tags: [gate-passes]
      summary: Get a gate pass by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gate pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePass"
        "404":
          description: Not found
  /gate-passes/{id}/status:
    patch:
      operationId: updateGatePassStatus
      tags: [gate-passes]
      summary: Approve or reject a gate pass (warden)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GatePassStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePass"
  /gate-passes/{id}/mark-out:
    patch:
      operationId: markGatePassOut
      tags: [gate-passes]
      summary: Mark student as exited campus (security)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePass"
  /gate-passes/{id}/mark-in:
    patch:
      operationId: markGatePassIn
      tags: [gate-passes]
      summary: Mark student as returned to campus (security)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePass"
  /gate-passes/{id}/qr:
    get:
      operationId: getGatePassQr
      tags: [gate-passes]
      summary: Get QR code data for a gate pass
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeResponse"
  /gate-passes/verify/{token}:
    get:
      operationId: verifyGatePass
      tags: [gate-passes]
      summary: Verify a gate pass by QR token (security)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Valid pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GatePass"
        "404":
          description: Invalid or expired
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, warden, security, admin]
        rollNumber:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        hostelId:
          type: ["integer", "null"]
        departmentId:
          type: ["integer", "null"]
        roomNumber:
          type: ["string", "null"]
        parentPhone:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, warden, security, admin]
        rollNumber:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        roomNumber:
          type: ["string", "null"]
        parentPhone:
          type: ["string", "null"]
        hostelId:
          type: ["integer", "null"]
        hostelName:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    GatePass:
      type: object
      required: [id, studentId, outingType, reason, destination, outgoingTime, expectedReturnTime, status, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: ["string", "null"]
        studentRollNumber:
          type: ["string", "null"]
        hostelName:
          type: ["string", "null"]
        roomNumber:
          type: ["string", "null"]
        outingType:
          type: string
          enum: [home_visit, weekend_leave, emergency_leave, medical_leave]
        reason:
          type: string
        destination:
          type: string
        outgoingTime:
          type: string
          format: date-time
        expectedReturnTime:
          type: string
          format: date-time
        actualOutTime:
          type: ["string", "null"]
          format: date-time
        actualInTime:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [pending, approved, rejected, out, returned]
        wardenRemarks:
          type: ["string", "null"]
        approvedById:
          type: ["integer", "null"]
        approvedByName:
          type: ["string", "null"]
        qrToken:
          type: ["string", "null"]
        isLate:
          type: boolean
        createdAt:
          type: string
          format: date-time
    GatePassInput:
      type: object
      required: [outingType, reason, destination, outgoingTime, expectedReturnTime]
      properties:
        outingType:
          type: string
          enum: [home_visit, weekend_leave, emergency_leave, medical_leave]
        reason:
          type: string
        destination:
          type: string
        outgoingTime:
          type: string
          format: date-time
        expectedReturnTime:
          type: string
          format: date-time
    GatePassStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        wardenRemarks:
          type: ["string", "null"]
    GatePassListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/GatePass"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    QrCodeResponse:
      type: object
      required: [qrDataUrl, token]
      properties:
        qrDataUrl:
          type: string
        token:
          type: string
...
    Department:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        studentCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
...