# lib/api-zod/src/generated/types/gatePassInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Campus GatePass System API
 * OpenAPI spec version: 0.1.0
 */
import type { GatePassInputOutingType } from './gatePassInputOutingType';

export interface GatePassInput {
  outingType: GatePassInputOutingType;
  reason: string;
  destination: string;
  outgoingTime: Date;
  expectedReturnTime: Date;
}