# lib/api-zod/src/generated/types/gatePassStatusUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Campus GatePass System API
 * OpenAPI spec version: 0.1.0
 */
import type { GatePassStatusUpdateStatus } from './gatePassStatusUpdateStatus';

export interface GatePassStatusUpdate {
  status: GatePassStatusUpdateStatus;
  /** @nullable */
  wardenRemarks?: string | null;
}