# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Campus GatePass System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  rollNumber?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  roomNumber?: string | null;
  /** @nullable */
  parentPhone?: string | null;
  /** @nullable */
  hostelId?: number | null;
  /** @nullable */
  hostelName?: string | null;
  /** @nullable */
  departmentId?: number | null;
  /** @nullable */
  departmentName?: string | null;
  isActive?: boolean;
  createdAt: Date;
}